# api/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "KOBE-Tech API documentation generated from Go annotations.",
        "title": "KOBE-Tech API",
        "contact": {},
        "version": "0.1.0"
    },
    "basePath": "/",
    "paths": {
        "/api/articles": {
            "get": {
                "description": "Get article list API.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "articles"
                ],
                "summary": "List articles",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.listArticlesResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/server.articleErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create article API.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "articles"
                ],
                "summary": "Create article",
                "parameters": [
                    {
                        "description": "Create article request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/server.createArticleRequest"
                        }
                    }
                ],
                "responses": {
                    "501": {
                        "description": "Not Implemented",
                        "schema": {
                            "$ref": "#/definitions/server.notImplementedResponse"
                        }
                    }
                }
            }
        },
        "/api/auth/login": {
            "post": {
                "description": "Login API.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Login",
                "parameters": [
                    {
                        "description": "Login request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/server.loginRequest"
                        }
                    }
                ],
                "responses": {
                    "501": {
                        "description": "Not Implemented",
                        "schema": {
                            "$ref": "#/definitions/server.notImplementedResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "article.ArticleJSON": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string",
                    "example": "Article body"
                },
                "created_at": {
                    "type": "string",
                    "example": "2026-04-24T00:00:00Z"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "title": {
                    "type": "string",
                    "example": "First article"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2026-04-24T00:00:00Z"
                },
                "user_id": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "server.articleErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "internal server error"
                }
            }
        },
        "server.createArticleRequest": {
            "type": "object",
            "properties": {
                "body": {
                    "type": "string",
                    "example": "Article body"
                },
                "title": {
                    "type": "string",
                    "example": "First article"
                }
            }
        },
        "server.listArticlesResponse": {
            "type": "object",
            "properties": {
                "articles": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/article.ArticleJSON"
                    }
                }
            }
        },
        "server.loginRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "format": "email",
                    "example": "user@example.com"
                },
                "password": {
                    "type": "string",
                    "format": "password",
                    "example": "change-me"
                }
            }
        },
        "server.notImplementedResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "feature is not implemented yet"
                },
                "next_step": {
                    "type": "string",
                    "example": "internal/{domain}/{handler,service,repository}.go"
                }
            }
        }
    }
}